// apps/Main/src/pages/COMM/COMM010E03/layout/Types.ts
export type FormDataType = {
  user_name: string;
  user_id: string;
  pwd: string;
  pwd_new: string;
  pwd_chk: string;
  busi_place_name: string;
  emp_no: string;
  user_level: string;
  unit_work_name: string;
  plant_name: string;
  equip_name: string;
  responsi_name: string;
  email: string;
  mobile_no: string;
  use_yn: string;
  remark: string;
  row_stat?: 'modified';
};